rename code block and move file to satisfy prettier
