+++ content/getting-started/quick-start.mdx
@@ -15,7 +15,7 @@ First, [create a Knock account](https://dashboard.knock.app/signup) if you don't
 
 We have SDKs available in [most major languages](/reference#client-libraries). Don't see your language listed here? [Let us know](mailto:support@knock.app)!
 
-<MultiLangCodeBlock snippet="install" title="Install Knock" />
+<MultiLangCodeBlock snippet="sdks.install" title="Install Knock" />
 
 ## 2. Set up environment variables
